requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
configparser==6.0.0
dataclasses-json==0.6.2
exceptiongroup==1.1.3
frozenlist==1.4.0
gigachat==0.1.8
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.336
langsmith==0.0.64
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pydantic==2.5.1
pydantic_core==2.14.3
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
telebot==0.0.5
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.2
